The root cause is the pre-registration of the triple store in the global row store before the axioms and vocabulary objects have been initialized.  Thus, a concurrent attempt to locate the triple store can succeed and return the row store properties without the axioms or vocabulary objects.
This issue is clearly demonstrated by TestConcurrentKBCreate.

The root cause was the GRS was being written in super.create().  It was then being updated after the lexicon and spo relations were initialized.  So, a discovery would succeed since it was already in the GRS but the vocabulary and axioms had not yet been initialized.  I fixed this by passing in the container to the LexiconRelation and SPORelation constructors and deferring the GRS writes until after everything was fully initialized.  I also move the code to publish the reference into the default resource locator cache until after the commit() in create().

TestLocalTripleStore, the AST Evaluation test suite, TestConcurrentKBCreate(), and TestBigdataSailWithQuads are all green with this change.

@see https://sourceforge.net/apps/trac/bigdata/ticket/617 (Concurrent KB create fails with "No axioms defined?")

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6707 8f7bc0f5-282e-0410-95e3-8d296e9bb460