I am reverting this change in order to have a direct apples to apples test.
Another possible explanation for reduced throughput is that we are leaking some Futures with the consequence that some B+Tree nodes and leaves are being pinned in memory - more than are accounted for by the 4000 capacity write retention queue.  That hypothesis could be testing simply by changing the capacity of the write retention queue.  This is part of the index metadata, so changing it requires an update to the index declaration or a reload.

Unfortunately we do not report the index performance counters for anything other than the unisolated index view.  We should consider using CATs and updating those counters for query by sharing the counters across either ALL views or having one set of counters for the unisolated views and another set for all read-only views.

@see https://sourceforge.net/apps/trac/bigdata/ticket/625 (Possible failure to clear Future from Memoizer cache for BTree/HTree)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6758 8f7bc0f5-282e-0410-95e3-8d296e9bb460