Added unit test for [1] from the ticket.
Note: The test is not yet integrated into CI since it is not yet known whether the test will pass once the NotMaterializedException problem has been resolved.

The NotMaterializedException is being through due to a getValue() call without a proceeding setValue() call on an inline IV (XSD Boolean in this case).  The logic that sets the IV Cache needs to be checked to decide whether we want to set the IV Cache for inline IVs or make the code invoke asValue() rather than getValue() when the IV is inline.

[1] https://sourceforge.net/apps/trac/bigdata/ticket/632 (NotMaterializedException in SERVICE call)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6813 8f7bc0f5-282e-0410-95e3-8d296e9bb460