IAtomicStore and IBufferStrategy now declare an isDirty() method. The semantics of this method is that is reports true iff there has been a write on the store/strategy since the last commit() or abort().
One change with operational semantics was also introduced.  nextOffset is reset to the last committed value of nextOffset by abort() for the WORM style IBufferStrategy implementations.  This appears to be an oversight leading to wasted space when handling an abort in historical versions of the code base.

BigdataSailConnection#close() was modified to (a) recognize when updates made through the sail had been flushed from the assertion and/or retraction buffers to the various indices - this is captured by a new BigdataSailConnection.isDirty() method; and (b) to automatically invoke rollback() when a BigdataSailConnection was closed and isDirty() was true.  This change should protect people who are not using the correct pattern to guarantee invocation of rollback() in the error handling code for an unisolated connection.  (Failure to invoke rollback for an unisolated connection would leave dirty index pages in memory and cause them to be flushed through to the disk with the next commit, resulting in a corruption of the index data structures).

@see https://sourceforge.net/apps/trac/bigdata/ticket/640 (Conditionally rollback() BigdataSailConnection if dirty)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6885 8f7bc0f5-282e-0410-95e3-8d296e9bb460