Refactored the HALog file extension onto the IHALogReader interface (in both the halog and althalog packages).  This data was being declared in several different locations.
Added declaration for the quorum/backup znode to ZKQuorum in support of HA backups.

Added QuorumBackupState object in support of HA backups.

Modified purgeHALogs() to check the data on the quorum/backup znode.  When that znode exists, it will not delete HALogs unless they have been backed up.

Added code to HAJournalServer.serviceJoin() to purge HALogs (but not the current HALog) when we transition to a fully met quorum.  This allows the purge to occur before we reach a commit point.  This code is conditionally disabled for now.

Additional cleanup and documentation on HAJournalServer and the HA transition flowcharts.

Removed the incremental option from IHASyncRequest.  This is no longer required since we have restructured how we do RESYNC.



git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/READ_CACHE@6924 8f7bc0f5-282e-0410-95e3-8d296e9bb460