Added method to report the pid of the HAJournalServer process out through the NSS status page.
Added javadoc concerning testStartABCSimultaneous() deadlock.  We now have stack traces that show what is going on here.  The problem is that the 3rd service is awaiting the set of the quourm token and installation of the root blocks from the leader while the leader is waiting on WriteCacheService.flush() which is waiting on the 3rd service to setup the pipeline (and hence waiting for the quorum token and root block install).


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/READ_CACHE@6969 8f7bc0f5-282e-0410-95e3-8d296e9bb460