Modified the IHAWriteMessage to include the UUID of the store on which the write was applied by the leader.  This is used in handleReplicatedWrite() to avoid invoking pipelineSetup() for a live write before the local root blocks are consistent with those of the leader. This change was introduced to address a deadlock problem in CI for the ABC "Simultaneous" start test.
Note: It was a bit messy to get the storeUUID into the RWStore due to the multiplicity of code paths in the handling of RWStore initialization in the ctor and in initFromRootBlock().

There was an undetected problem where the JournalMetadata was immutable, but in HA we can now replace the storeUUID with the UUID from the leader.  I updated the relevant code paths in AbstractJournal to handle this case (getUUID(), etc).

I have updated the HAJournalServer CI test worksheet.  All looks good.

I have run the various WCS, HA, Journal and SPARQL level test suites locally.  All looks good there too.



git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/READ_CACHE@6974 8f7bc0f5-282e-0410-95e3-8d296e9bb460