I have added IHABufferStrategy.writeOnStore().  The caller is responsible for taking a read-lock, opening the output stream, and then writing the root blocks onto the stream.  This method then writes the data (starting immediately after the root blocks) onto the stream.  The current approach writes the journal's data onto a GZIPOutputStream, thus the snapshot is compressed.
This is integrated into HAJournal using a temporary file rename pattern. The snapshot is written onto a temporary file in the snapshot directory.  Once the file has been written, flushed, and closed it is renamed onto the target snapshot file name.  If the process is interrupted, a temporary file will exist in the snapshot directory.  That temporary file may be removed.  Note that it is not necessary for this pattern for the file rename to be atomic.  If the file winds up renamed, then it is well and good.  If the file does not wind up renamed, then the snapshot was is invalid and will not be used since it will have the wrong name.

I have written implementations of this writeOnChannel() method for the WORMStrategy and RWStrategy.  These implementations have not yet been tested. 

We still need to write the test suites for the HA Backup / restore mechanisms.

See https://sourceforge.net/apps/trac/bigdata/ticket/645 (HA Backup)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/READ_CACHE@7010 8f7bc0f5-282e-0410-95e3-8d296e9bb460