Added a test suite for the HA backup scenarios.  This test suite is not yet complete.  It verifies the ability to create snapshots, but does not yet validate the snapshots or the HARestore utility.
Added HAGlue.takeSnapshot() method.  This is in addition to the REST API for requesting a snapshot.  

Moved the takeSnapshot() logic entirely inside of the SnapshotManager.

Added logic to the HA test suite harness to permit overrides of configuration parameters. This is used to selectively change the snapshotPolicy and retentionPolicy in the test suites which exercise that logic.

HA test suite is good (worksheet updated).

Reconciled an edit by MC for this commit.

@see https://sourceforge.net/apps/trac/bigdata/ticket/645 (HA Backup)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/READ_CACHE@7016 8f7bc0f5-282e-0410-95e3-8d296e9bb460