Changed the default #of write cache buffers for HAJournal.config from 12 to 500.  This provides a substantial reduction in the IO Wait associated with the leader and allows the use with traditional disk (SATA) as well as SAS and SSD.
Added a means to list out the existing snapshots and exposed that list
on the "GET .../status" page.

Added HAGlue method to compute the digest of a snapshot.

Bug fix to computeDigest() methods.  The implementations were using
digest.digest() rather than digest.update().  This means that the
digest was only valid for the last buffer worth of data!

Bug fix to RWStore.computeDigest(). This method was computing the
digest starting after the root blocks.  The digest methods should
compute the digest of the entire file.  I have also updated the
documentation to clarify this.

Added ?digests option to the "GET .../status" page.  This option will
print the digests for the Journal, HALogs, and snapshot files. This is
only a debugging tool.

Using the ?digests option, I was able to determine that the RWStore
and HALog digests were (still) not being computed correctly.

The snapshot MD5 digests agree with those computed by the md5 utility
in the OS.

The RWStore and WORMStrategy compute digest methods were broken.  They
were failing to update the offset after each block read.  These
methods were also modified to use the digest.update(ByteBuffer)
variant for better efficiency and the temporary byte[] was removed
from the code.

HALog md5 was also broken with the same root cause.  In addition, the
sequence counter was not being udpated and the ByteBuffer needed to be
flipped after the FileChannelUtility.readAll() call.  I fixed both
implementations.  They now provide the same results as the OS md5
utility.

Fixed a bug in RESTORE where an endless loop could occur if a service
had an empty HALog file lying around.  During RESTORE, the HALog files
are applied one by one for each commit point that is +1 over the
current committed data.  Since the opening and closing root blocks are
the same for an empty HALog file, this was failing to advance the
commit point on the journal and thus attempting to re-apply the same
file again and again.

No change in HA test suite failures.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/READ_CACHE@7023 8f7bc0f5-282e-0410-95e3-8d296e9bb460