Probable bug fix for testAB_halog(), testAB_C_halog(), and testStartABCSimultaneous().  These tests were failing to wait until the first commit point had become visible on each of the services.  I added a helper method for this: awaitCommitCounter().
I have removed all use of awaitKBCreate() and replaced it with
awaitCommitCounter(1,...).  The awaitCommitCounter(1,...) pattern is
also in the ABC() constructor for both sequential and concurrent
service starts.

I still observe occasional failures in bounce leader and bounce follower.  These may be due to transitions into an error state that are not cured.  We have some commented out code in HAJournalServer for this that still needs to be explored.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/READ_CACHE@7031 8f7bc0f5-282e-0410-95e3-8d296e9bb460