I have lifted out the JVMHashIndex.  This was an inner class of the JVMHashJoinUtility.  It is now a top-level class in the same package.
Refactored makeKey() by lifting in the hashCode computation and removing the thrown exception when a join variable was not bound.  This should reduce overhead associated with exception throws that could be handled through other means.  One side-effect is that the hashCode is no longer always ONE for a solution having an unbound variable.  Instead, we simply ignore that variable when computing the hash code but continue to compute the hash code based on the variables that are bound. Solutions with unbound variables are discovered during a post-join scan in which we identify the solutions that did not join.  Thus, this should even help with OPTIONAL, NOT EXISTS, or MINUS joins when there are a large number of solutions that did not join by distributing them throughout the buckets rather than always in the bucket associated with the hash code ONE.  

Rationalize the semantics of the "optional" and "joinVars" parameters for the JVMHashIndex.  The joinVars has been relabeled the keyVars and merely indicates which variables are used to form the keys for the hash index.  The optional flag indicates whether or not the key variables MUST be bound. When optional:=false, an unbound key variable is simply ignored.

Modified GroupNodeBase to output additional information (joinVars and projectInVars).

Extracted IDistinctFilter interface with both accept() and filterSolutions() methods. accept() is for solution at a time processing.  filterSolutions() supports vectored processing.  Modified JVMDistinctBindingSets and JVMDistinctFilter to use and support the new interface.  I have started work on HTree version of this interface. The HTree version will be used to impose a DISTINCT filter in the solutions projected into a sub-group. The HTree version is more complex because of the vectoring optimizations that we perform for the HTree.  There are various notes on this in HTreeHashJoinUtility.

All SPARQL and BOP related tests pass.  The DISTINCT filter option is disabled in the committed code.  One of the next steps is to determine whether the ASTComplexOptionalOptimizer is still required.  The specific customer issue which motivated this ticket is fixed primarily by disabling that optimizer (it reduces the run time to less than 1 second). 

The DISTINCT filter on the solutions flowing into the subgroup has much less of an impact on overall performance for that query.  However, the DISTINCT filter does provide a significant performance gain for some queries that we have extracted from the original query as illustrated below.

# With DISTINCT filter in outputSolutions():
# 
# Note: About 1/2 this if we use count(*) versus select(*)
#
# solutions=180953, chunks=1810, subqueries=0, elapsed=8997ms.
# solutions=180953, chunks=1810, subqueries=0, elapsed=8702ms.
#
# Without DISTINCT filter in outputSolutions():
#
# solutions=180953, chunks=1813, subqueries=0, elapsed=210856ms

These various issues are clearly worth continued pursuit.

See https://sourceforge.net/apps/trac/bigdata/ticket/668 (JoinGroup optimizations).

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@7089 8f7bc0f5-282e-0410-95e3-8d296e9bb460