Added a concept for an "Operator" run state.  Added some stubs for operator alerts.  This is only displayed through the /status page for now (through getExtendedRunState()).
Added a boolean option that may be used to enable the online disaster recovery (REBUILD).  It is disabled by default for now to avoid "accidental" rebuilds.  We could give people a button to press to start a rebuild on the status page, but right now I want to make sure that we do not start rebuilds unnecessarily.

Added override in TestHA3JournalServer to enable the testing of REBUILD even if it is disabled by default.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/READ_CACHE@7094 8f7bc0f5-282e-0410-95e3-8d296e9bb460