Checkpoint on refactor that will introduce an index over the HALog files to reduce the latency at the commit.
This abstracts out a common base class for the SnapshotIndex and the HALogIndex.  The HALogIndex is not yet in use.

The next step is to see if I can get the test suite to pass with the alt HALogManager class.

See https://sourceforge.net/apps/trac/bigdata/ticket/670 (Accumulating HALog files cause latency for HA commit)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/READ_CACHE@7113 8f7bc0f5-282e-0410-95e3-8d296e9bb460