AbstractJournal: Added logic to make AbstractJournal.commit() spin for up to 5000ms if required to obtain a commitTime that moves forward with respect to the then current lastCommitTime.  This addresses a potential source of problems with clock skew in HA and could permit us to allow more skew to be present among the nodes of an HA cluster (though I have not yet relaxed that constraint in Journal.assertBefore()).
AbstractJournal: added getRootBlockWithLock() variant.

HAJournalServer.awaitJournalToken(): modified to use getRootBlockWithLock() to ensure atomicity of the state changes that awaitJournalToken() is monitoring.

Journal: More cleanup on the consensus release time protocol.  Changes improve reporting of the root cause of an exception in the release time consensus protocol and shift reporting of clock synchronization problems entirely to the leader (based on the timestamps obtain on the leader and the followers).

@see https://sourceforge.net/apps/trac/bigdata/ticket/530 (Journal HA)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/READ_CACHE@7151 8f7bc0f5-282e-0410-95e3-8d296e9bb460