AST2BOpUtility - javadoc only.
TestInclude - added a 2nd test case so we have coverage both for the code path where we use a SCAN (INCLUDE runs first) as well as the case where we run a JOIN first and then join in the data from the INCLUDE.

include_03.rq - SPARQL comment changes only.

include03a.rq - the new version of the test for TestInclude.

BigdataSPARQLUpdateTest - javadoc only.

BigdataSPARQLUpdateTestTx - javadoc only.

@see https://sourceforge.net/apps/trac/bigdata/ticket/531 (SPARQL UPDATE for NAMED SOLUTION SETS)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@7167 8f7bc0f5-282e-0410-95e3-8d296e9bb460