Refactored the HAReceiveService to be more robust to pipeline change events.  This is to address an issue that Martyn has observed in HA CI runs where a CancellationException would be caught in QuorumPipelineImpl. The CancellationException is being interpreted as a normal termination, but what we want is the pipeline change event to trigger a retrySend(). This was observed for the changeDownstream event.
The change to the HAReceiveService is that we no longer interrupt the ReadTask (via the ReadFuture) in either changeDownstream() or changeUpstream(). Instead, we set a firstCause Throwable. If that firstCause is non-null, then it is thrown out of the ReadTask and the socketChannel is closed. This causes a well known exception (PipelineDownstreamChange or PipelineUpstreamChange) to be propagated back to retrySend().  We do not need to catch these exceptions explicitly since retrySend() will catch everything.  However, the pipeline change now no longer causes a CancellationException to be propagated and thus retrySend() now will actually retry the replicated write.

@see https://sourceforge.net/apps/trac/bigdata/ticket/681 (Journal HA)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/READ_CACHE@7190 8f7bc0f5-282e-0410-95e3-8d296e9bb460