I have modified HALogNexus to NOT add an empty HALog file on startup to the in-memory index iff it is the last HALog file in commit counter sequence for that service.
I have written a unit test for this behavior (TestHA3JournalServerWithHALogs.test_startABC_emptyLogFileDeletedOnRestartC()).  I have verified that the test replicates the problem reported here and that the change to HALogNexus causes the test to run green.

I found and fixed a bug in the test harness where copyFiles() would not actually copy a single file (the fix was an update to the documentation).  I also exposed the copyFile() method to subclasses.

Note: This commit does not include the test suite.  I need to vet some other edits in the test suite before I can commit that code. However, this commit does fix the problem desribed for this ticket.

@see http://sourceforge.net/apps/trac/bigdata/ticket/679 (HAJournalServer can not restart due to logically empty log file)


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/READ_CACHE@7197 8f7bc0f5-282e-0410-95e3-8d296e9bb460