The failure to wake the producer when you call buffer.iterator().close() is because you are not calling buffer.setFuture() with the Future of the producer in your test.  I have committed a test which does this and shows that the producer is correctly awoken (the test observes a CancellationException for the Future).
Per the ticket, please verify under a debugger that the iterator() and/or buffer were closed for your stack traces.  I suspect that they were not closed and that the traces exist because the query was not correctly terminated.  It is also possible that setFuture() was not called on the buffer or perhaps that a race condition existing for setting the Future on the buffer.  However, I need more information to diagnose this furthre.

See https://sourceforge.net/apps/trac/bigdata/ticket/707 (BlockingBuffer.close() does not unblock threads)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@7236 8f7bc0f5-282e-0410-95e3-8d296e9bb460