It turns out that the variable <code>now</code> in the GatherTask was only used for the error response.  The method {{{         protected IHANotifyReleaseTimeRequest newHANotifyReleaseTimeRequest(final UUID serviceId) }}} was actually taking its own timestamp using the following method, which is designed to allow override from the HA test suite. {{{ private long newConsensusProtocolTimestamp() }}}
Since the clock skew issue arose only in the error path, I have modified the error path to use nextTimestamp() and thus report a reasonable timestamp and avoid this false clock skew error.

I have also modified one of the HA tests to use the ABC() utility to start of the services.  It was replicating that logic inside of the test code.

@See #720

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/READ_CACHE2@7280 8f7bc0f5-282e-0410-95e3-8d296e9bb460