Refactored to add link weight support (still need the RDR parser to use this except in SIDs mode and still need to co-locate that SIDs with the ground triples in the SPO(C) and OSP(C) indices).
Refactored nearly everything to abstract out a concrete implementation that does not depend on the bigdata backend.  The next step will be to replace the use of IV and ISPO throughput with openrdf Value and openrdf Statement.  We can then easily implement a variant over the MemorySail for openrdf (or over some custom classes that are optimized for the GAS traversal patterns).  That will give us a fully functioning implementation outside of the core bigdata code base.  We can then isolate the remaining dependencies (striterator and a few other items) and create an Apache 2 package for the APIs and the non-bigdata specific implementation.

See #629

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@7346 8f7bc0f5-282e-0410-95e3-8d296e9bb460