Added a managed array abstraction and replaced the static frontier implementation with that abstraction.
The code paths that compact and sort the frontier have been modified to reuse an array that grows as necessary.  This should reduce the overhead associated with memory allocation for the frontier.

I still need to optimize this for the thread-local frontier.

I still need to use a parallel sort for the frontiers (other than the thread-local one, which already breaks the sort into N threads since each per-thread frontier is sorted independently).

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@7353 8f7bc0f5-282e-0410-95e3-8d296e9bb460