Bug fix for #708 (BIND heisenbug)
The root cause was a failure to clone the binding set before modifying it.

---

This view of the same query run state log data (for the same run) shows the problem:
{{{
step	label	bopId	fanIO	navail	nrun	cnksIn	unitsIn	cnksOut	unitsOut
1	startQ	4	1	1	0				
2	startOp	4	1	0	1				
3	startOp	3	1	-1	2				
4	startOp	2	1	-2	3				
5	startOp	6	1	-3	4				
6	startOp	7	1	-4	5				
7	haltOp	4	2	-2	4	1	1	2	2
8	haltOp	2	1	-1	3	1	1	1	1
9	haltOp	3	1	0	2	1	1	1	1
10	startOp	1	1	-1	3				
11	haltOp	7	0	-1	2	1	1	0	0  <=== NOTHING OUT
12	haltOp	6	1	0	1	1	1	1	1
13	startOp	8	1	-1	2				
14	haltOp	1	1	0	1	1	1	1	1
15	haltOp	8	0	0	0	1	1	1	1
}}}

BopId 7 is one of the BIND operators.  In the table, you can see that it did not output any chunks before it halted.  Given that nothing was output, there was only one solution flowing into bop#7 (the copyOp where the two solution paths were combined) and only one solution flowing into bop#8 (the projection operator).  

This is conclusive evidence that the problem is not with the RunState accounting.  The HaltOp message which was generated for the BIND() clearly indicates that there was no solution produced.  Perhaps the solution was eliminated by the BIND()?  E.g., due to an interaction through the binding set objects?

Ok. The fix is to ConditionalRoutingOp#171. Replace
{{{
                        final IBindingSet bset = chunk[i];
}}}
with
{{{
                        final IBindingSet bset = chunk[i].clone();
}}}

This cures the test failures.

I have also added source.close() to the finally{} clause, but this is not required in order to cure the problems.

com.bigdata.rdf.sparql.ast.eval.TestAll is good.

I am going to assign the ticket back to Mike to take a look at the code in ConditionalRoutingOp and see whether it can be optimized with in the light of the need for this clone().


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@7375 8f7bc0f5-282e-0410-95e3-8d296e9bb460