I've added a simple InferenceChangeLogReporter. This class may be used to obtain the actual inferences computed during a transaction.
I modified the change log listener test suite to also test this this new listener.

The present implementation uses a LinkedHashMap to store the ISPOs for the inferences.  This should be scalable up to millions of inferences.  If very large inferences will be drawn, then we could substitute an HTree index for the LinkedHashMap. The existing NativeDistinctFilter already does exactly this and simply replace the hard coded use of the LinkedHashSet for improved scaling.  In this case, the native memory associated with the HTree needs to be released, e.g., through an ICloseable protocol on the change listener.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_3_0@7389 8f7bc0f5-282e-0410-95e3-8d296e9bb460