Rolling back to zk 3.3.3.  zk 3.4.5 has a reverse DNS lookup that causes a 5 second lag in the test case runs and throws off the test suite.
Added a private LookupDiscoveryManager to AbstractServer. This is used to export the service proxy rather than relying on the LookupDiscoveryManager in the HAClient.  This allows us to leave the service proxy exported until the tear down.

AbstractJournal.setQuorumToken2() modified to be robust when the HAQuorumService is terminated.

BigdataServlet modified to log if HTTP request was denied for HA NotReady or not Leader.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/ZK_DISCONNECT_HANDLING@7475 8f7bc0f5-282e-0410-95e3-8d296e9bb460