I have added an init.d style script:
bigdataHA (start|stop|status|restart)

This script relies on a bigdataHAEnv that MUST define the following variables to specify the location of the installed scripts. These variables SHOULD use absolute path names.

binDir
pidFile

This script could be used by an rpm or other installer to install the HA replication cluster as an init.d style service on a linux platform.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_3_0@7585 8f7bc0f5-282e-0410-95e3-8d296e9bb460