Sync to Martyn and CI on #724 (write replication pipeline resynchronization).
We have incorporated logic to drain to the marker in the replication protocol.

The marker concept has been refactored. There is now an IHASendState that captures original and potentially routing information for the payload.  This has been raised into the HAPipelineGlue interface.  

We still need to bring in the typed exception handling for forceRemoveService() invocations and examine QuorumPipelineImpl for possible lock contention issues around this resync protocol.

Removed dead test suite from build.xml

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/MGC_1_3_0@7636 8f7bc0f5-282e-0410-95e3-8d296e9bb460