The resetPipeline() code now takes the outer lock and then tears down the send/receive service and restarts them.  This does not clear up the problem, especially when the HASendService is sending 50k packets.
I tried replacing the sendLock with a semaphore to get more confidence that were not yielding the lock and allowing more than one send at a time.  This does not change anything.

We need to drill down further on the behavior when the send/receive services are torn down and the setup again and how this is interacting with the timing of the forceRemoveService() and the re-replication of the message and payload from the leader.  I have not yet drilled down on the logs in any depth for this version of the code.  I expect that there is an interaction between the timing of those different activities such that we get a repeated series of resetPipeline() requests rather than actually recovering a pipeline in a known good order and clean slate condition.

See #724.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/MGC_1_3_0@7672 8f7bc0f5-282e-0410-95e3-8d296e9bb460