Reduced the "explain" statistics to a simple table and added an "explain=details" option to provide all of the available detail. The cluster oriented statistics are no longer displayed unless you are running a scale-out cluster.  The mutation statistics are no longer displayed for queries (they really only apply to inference rules).
The simple view has the following columns:
queryId	deadline	elapsed	cause	evalOrder	bopSummary	predSummary	nvars	fastRangeCount	sumMillis	unitsIn	unitsOut	typeErrors	joinRatio

In fact, this really should be replaced by parameterized styling of the XHTML result page to render the appropriate level of detail for the user rather than doing this in the server side of the code.  When we do this, we can just specify detailedStats=clusterStats=mutationStats=true to get all the data into the page and then style the page to render only the bits that are relevant to the user.



git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_3_0@7689 8f7bc0f5-282e-0410-95e3-8d296e9bb460