The RTO now supports FILTERs on required joins when those filters have materialization requirements.  This is the subject of #257, which is now closed.  Over time we will probably refactor the RTO to operate directly on the AST nodes (e.g., StatementPatternNode versus SPOPredicate), but it currently handles conditional materialization pipelines just fine.
There are currently two different ways in which a cutoff join can be evaluated.  One uses a single pipeline join and carefully controls the execution of that join to obtain an estimated cardinality.  This is the historical method.

The other can handle a sequence of operators.  A pipeline join is generated followed by whatever operators are required to materialize variable bindings and evaluate filters. Finally, a SLICE is appended to the query plan to limit the output.  When the query is executed, a rowid column is injected into the source solutions. This is used to correctly correlate the #of input solutions required to produce a given #of output solutions and hence obtain the estimated cardinality of the join through cutoff evaluation.

Some slight differences in the resulting plans and runtime behavior have been observed when all queries 

See #257 (Support BOP fragments in the RTO)
See #64 (Runtime Query Optimizer)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_3_0@7755 8f7bc0f5-282e-0410-95e3-8d296e9bb460