The fix for (1) is easier than I had anticipated.  The problem was in {{{ ASTSparql11SubqueryOptimizer::     private void liftSparql11Subquery(final AST2BOpContext context,             final StaticAnalysis sa, final SubqueryRoot subqueryRoot) { }}} This method was failing to copy across the attached join filters to the new NamedSubqueryInclude. {{{         /**          * Copy across attached join filters.          *          * @see <a href="https://sourceforge.net/apps/trac/bigdata/ticket/796"          *      >Filter assigned to sub-query by query generator is dropped from          *      evaluation</a>          */         include.setAttachedJoinFilters(subqueryRoot.getAttachedJoinFilters()); }}}
This change fixes both the case where the FILTER requires materialization of RDF Values and the case where it does not (which was the case for (1)).

This fix also highlighted three unit tests where we had failed to test for the correct attachment of the FILTER to the lifted out NamedSubqueryInclude.

TestASTSparql11SubqueryOptimizer::
* test_subSelectWithLimitAndOrderBy
* test_subSelectWithAggregation
* test_subSelectWithRunOnceAnnotation

Regression tests for (1) are captured in the TestSubQuery class.

Finally, I do observe that the join filters are attached to the SPARQL Sub-Select when it is not lifted out.



git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_3_0@7804 8f7bc0f5-282e-0410-95e3-8d296e9bb460