Resolution for #803 (prefixMatch does not work in full text search).
The root cause was a change in how this feature is invoked.  It is now indicated by a wildcard in the bsd:search literal.

Added a unit test for the simplified version of this syntax.

Cleaned up many references to bd:search. The default namespace for this is now bds.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_3_0@7814 8f7bc0f5-282e-0410-95e3-8d296e9bb460