I have modified HALogNexus to release the HALog files asynchronously.  This causes some problems in the existing unit tests, so I have added a configuration parameter that may be used to specify a synchronous HALog release timeout - this defaults to ZERO, but may be increased as desired (in exchange for latency in 2-phase commits) or as necessary for the existing tests, some of which assume that HALog files are released synchronously at the 2-phase commit.
The expected behavior should now be that HALog files are asynchronously release starting with the first 2-phase commit in which the corresponding commit point is reclaimable.  The release of HALog files will continue until the specified first commit point that may not be released, or until an invariant is violated.  Under sustained commits with concurrent releases, any commit in which the task to release HALog files is longer running will start a new such task with a new earliest commit point that may not be released.

Changes are to:

* HALogNexus: now submits a task to release the HALog files.  There can be at most one instance of this task running for a given HAJournal.  The task is wrapped by the FutureTaskInvariantMon and will be cancelled if the service is no longer joined with the met quorum.  

* HAJournalServer: A new configuration parameter exists to control the maximum delay for releasing the HALog files (default is ZERO milliseconds).

* TestHARestorePolicy: This was modified to raise the timeout for the synchronous purge of the HALogs so the test assumptions would remain valid.

Note: The HAJournalServer.ConfigurationOptions.HA_LOG_PURGE_TIMEOUT defaults to zero.  This is probably what we always want. However, the existence of this option allows us to revert to the old behavior using a configuration change or by changing the default.

See #780

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_3_0@7816 8f7bc0f5-282e-0410-95e3-8d296e9bb460