Modified the GangliaLBSPolicy to explicitly check for a null hostname (can be caused by a failed RMI).
Modified the HA status page to include the proxy object for the RMI interface for the local service.  I want to use this to diagnose situations where one service as a proxy for another but the proxy that is has does not respond and is, hence, probably dead.  Note that moving into an error state on that service whose exported proxy is not responding is not enough to get the service to become responsive.  This is probably because the error state does not unexport the proxy. However, it does not explain why/how a bad proxy got out there in the first place.

See #624 (HA LBS)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_3_0@8195 8f7bc0f5-282e-0410-95e3-8d296e9bb460