Working on #939 (NSS does not start from command line: bigdata-war/src not found).
NanoSparqlServer: code has been modified to explicitly search (if jetty.resourceBase is not defined) (a) the local file system; (b) the classpath; and then (c) default to whatever is the default value in jetty.xml for the jetty.resourceBase property.

TestNSSHealthCheck: added a basic test suite for checking the health of an NSS instance once deployed.  This is a starting point for CI based tests of the various deployment models.

build.xml: modified to illustrate a possible way of performing the CI deployment tests.  More needs to be done here!


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_3_0@8351 8f7bc0f5-282e-0410-95e3-8d296e9bb460