Bug fix for #940 (HA LBS breaks tomcat deployment).
The root cause is that the ProxyServlet is not available under tomcat (or anything else besides jetty).  Therefore it can not be configured from the same web.xml file that is used for other platforms.

To address this, I extracted the HA LBS configuration into a new override-web.xml file and then modified the NanoSparqlServer to locate that resource.

The HA test suite also needed to be modified to explictly locate this resource.

See #940 (HA LBS breaks tomcat deployment).

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_3_0@8359 8f7bc0f5-282e-0410-95e3-8d296e9bb460