See #988 (EXISTS is slow)
As indicated off-list, we have existing code paths that support a non-vectored subquery per source solution for (NOT) EXISTS using the SubqueryOp.

I have reenabled this code and verified that it passes the test suite (except for one detailed check of the physical operator plan that is generated for NOT EXISTS).

I have also added a LIMIT ONE when using the sub-query version.

However, I observe that the run time for the test described above is a constant .18 seconds regardless of which code path is used.  In neither case do I observe the slow performance described on this ticket of 5 seconds.  That is, the test case does not demonstrate the performance problem.  Please check the committed test cases in TestNegation.java.  See the test_exists_988a() and test_exists_988b() methods at the end of that file.

The sub-query plan is currently OFF
{{{
    private static PipelineOp addExistsSubquery(PipelineOp left,
            final SubqueryRoot subqueryRoot, final Set<IVariable<?>> doneSet,
            final AST2BOpContext ctx) {

        if (true) {
            // Vectored sub-plan evaluation.
            return addExistsSubqueryFast(left, subqueryRoot, doneSet, ctx);
        } else {
            // Non-vectored sub-query evaluation.
            return addExistsSubquerySubquery(left, subqueryRoot, doneSet, ctx);
        }
        
    }
}}}

I have not tested the use of a DISTINCT solutions operator in the sub-plan, but I have marked the code for where that operator should be introduced.
{{{
         * FIXME EXISTS: Try DISTINCT in the sub-plan and compare to correctness
         * without for (NOT) EXISTS and to performance of the non-vectored code
         * path for EXISTS>
}}}



git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_3_0@8524 8f7bc0f5-282e-0410-95e3-8d296e9bb460