Bug fixes to the test suite for #988.  I spoke with Jeremy and am now able to replicate the observed performance issue.  I have also verified that changing from the vectored sub-plan to the subquery LIMIT ONE approach fixes the performance problem for this query.
The next step is to create a query hint to direct the query plan generator to use the appropriate plan.

After that we can figure out if we can make the right decision automatically and also take a look at using DISTINCT over the variables in the FILTER in the vectored sub-plan to accelerate that code path.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_3_0@8525 8f7bc0f5-282e-0410-95e3-8d296e9bb460