Added query hint "filterExists" to control the behavior. The default is still the VectoredSubPlan.  You can specify the SubQueryLimitOne behavior like this: {{{ prefix eg: <http://www.bigdata.com/>
SELECT ?t
FROM eg:g
{ BIND (1 as ?t)
  FILTER EXISTS {
    ?a eg:p/eg:p/eg:p/eg:p/eg:p/eg:p/eg:p/eg:p ?b .
    hint:SubQuery hint:filterExists "SubQueryLimitOne" . # <<< OVERRIDE THE DEFAULT BEHAVIOR.
 }
}}}}

The unit test now verifies that the execution time is within target in order to verify that the query hint was correctly interpreted.

I had to modify the ASTQueryHintOptimizer to flow the QueryRoot into the IQueryHint handler method.  This caused all of the IQueryHint implementations to be touched.  I also had to expose the method on StaticAnalysis to locate the FILTER for a join group by entering in through the QueryRoot.

I fixed the problem where the ASTQueryHintOptimizer was not interpreting query hints in a FILTER.

I have not tried to introduce the DISTINCT SOLUTIONS into the VectoredSubPlan.

The core AST evaluation test suite is green.

See #988 (FILTER (NOT) EXISTS optimization)
See #990 (Query hint not recognized in FILTER)


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_3_0@8526 8f7bc0f5-282e-0410-95e3-8d296e9bb460