Several changes related to the single-server jetty NSS deployment model (startNSS).
build.xml: added comments about the purpose of the different log4j config files.

bigdata-war/classes/log4j.properties: modified to NOT specify the FileAppender for the rules log so as to avoid forcing the creation of rules.log file. This tends to be problematic for people since the default directory often does not have sufficient permissions.

startNSS: 
- Modified to fail if BD_HOME is not specified (and reports message).
- Wrapped many environment variables in quotes to avoid problems when they are empty.  This probably does not handle all edge cases with embedded whitespace yet.
- Modified to report the command that is executed and the PID that results.
- Modified to not send all output to /dev/null except for brew (the brew code path is probably incorrect, but I have kept the default behavior for now).


Note: The bigdataNSS script can not be used to control the run state of the NSS process because it has some brew metavariables.  

See #918 (Unify the Bigdata Installation Trees)


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_3_0@8564 8f7bc0f5-282e-0410-95e3-8d296e9bb460