Checkpoint on GIST refactor (#585) in support of group commit for the REST API (#566).
This commit introduces hierarchical locking (vs enumeration of the indices) into the AbstractApiTask and makes some progress toward a refactor of the ITask interface and the AbstractTask implementation to support access at the ICheckpointProtocol layer (supports BTree, HTree and Stream) in addition to the IIndex layer (supports BTree, FusedView, and IsolatedFusedView).

The next step will be to refactor AbstractTask.getIndex() to push down a method to obtaining the ICheckpointProtocol object.  In order to do this, the AbstractTask.indexCache must be relayered.  It is currently specific to the IIndex interface.  In order to provide caching for ICheckpointProtocol objects, it needs to be rewritten to the ICheckpointProtocol layer. However, the code needs to be carefully reviewed to determine whether we also need caching at the IIndex layer or if we can only cache at the ICheckpointProtocol layer.  This is a question of both correctness and performance.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_3_0@8578 8f7bc0f5-282e-0410-95e3-8d296e9bb460