This has been traced to a query hint override to control the operator parallelism (maxParallel=2).  While the HashIndexOp does c
Workaround: Bind the maxParallel query hint more closely to a specific join or other operator (sometimes difficult or impossible).

Workaround: Do not use the maxParallel query hint if the query has solution set hash joins (which appear if there is a nested complex optional join group, sub-select, INCLUDE, etc.)

Fix: Added ISingleThreadedOp interface and modified PipelineOp.getMaxParallel() to test for that interface and ignore annotation when the interface is present on an operator.  This interface was added to the following operators:

- HashIndexOp
- HTreeNamedSubqueryOp
- JVMNamedSubqueryOp
- HTreeHashJoinOp
- JVMHashJoinOp
- HTreeDistinctBindingSetsOp (also fixed unit tests that looked for UnsupportedOperationException rather than IllegalArgumentException)
- PipelinedAggregationOp (also fixed unit tests that looked for UnsupportedOperationException rather than IllegalArgumentException)

Other changes to:

- PipelineOp (added interface test and assertMaxParallelOne() method).

Note: Also reduced the #of threads for StressTestConcurrentUnisolatedIndices since it was failing in CI with "unable to create new native thread".

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_3_0@8619 8f7bc0f5-282e-0410-95e3-8d296e9bb460